use last args, script command shorthand
When inputting a command without any arguments, it uses the last valid arguments, saving typing when using multiple commands. Script command can now also be run using "s".
